Double precision (64-bit) LADSPA processing (needs special LADSPA filters to work, use at own risk)

+++ libecasound/ladspa.h
@@ -81,7 +81,7 @@ extern "C" {
    For audio it is generally assumed that 1.0f is the `0dB' reference
    amplitude and is a `normal' signal level. */
 
-typedef float LADSPA_Data;
+typedef double LADSPA_Data;
 
 /*****************************************************************************/
